moves upgrade files to cluster pkg
