+++ api/pkg/handler/routes_v1.go
@@ -596,7 +596,7 @@ func (r Routing) getMasterVersions() http.Handler {
 		endpoint.Chain(
 			r.oidcAuthenticator.Verifier(),
 			middleware.UserSaver(r.userProvider),
-		)(getMasterVersions(r.updateManager)),
+		)(cluster.GetMasterVersionsEndpoint(r.updateManager)),
 		decodeEmptyReq,
 		EncodeJSON,
 		r.defaultServerOptions()...,
@@ -1226,7 +1226,7 @@ func (r Routing) getClusterUpgrades() http.Handler {
 			middleware.UserSaver(r.userProvider),
 			middleware.Datacenter(r.clusterProviders, r.datacenters),
 			middleware.UserInfo(r.userProjectMapper),
-		)(getClusterUpgrades(r.updateManager, r.projectProvider)),
+		)(cluster.GetUpgradesEndpoint(r.updateManager, r.projectProvider)),
 		common.DecodeGetClusterReq,
 		EncodeJSON,
 		r.defaultServerOptions()...,
@@ -1252,7 +1252,7 @@ func (r Routing) getClusterNodeUpgrades() http.Handler {
 			middleware.UserSaver(r.userProvider),
 			middleware.Datacenter(r.clusterProviders, r.datacenters),
 			middleware.UserInfo(r.userProjectMapper),
-		)(getClusterNodeUpgrades(r.updateManager, r.projectProvider)),
+		)(cluster.GetNodeUpgradesEndpoint(r.updateManager, r.projectProvider)),
 		common.DecodeGetClusterReq,
 		EncodeJSON,
 		r.defaultServerOptions()...,
@@ -1278,8 +1278,8 @@ func (r Routing) upgradeClusterNodeDeployments() http.Handler {
 			middleware.UserSaver(r.userProvider),
 			middleware.Datacenter(r.clusterProviders, r.datacenters),
 			middleware.UserInfo(r.userProjectMapper),
-		)(upgradeClusterNodeDeployments(r.projectProvider)),
-		DecodeUpgradeClusterNodeDeploymentsReq,
+		)(cluster.UpgradeNodeDeploymentsEndpoint(r.projectProvider)),
+		cluster.DecodeUpgradeNodeDeploymentsReq,
 		EncodeJSON,
 		r.defaultServerOptions()...,
 	)
