+++ api/pkg/handler/routing.go
@@ -8,18 +8,9 @@ import (
 	prometheusapi "github.com/prometheus/client_golang/api"
 
 	"github.com/kubermatic/kubermatic/api/pkg/handler/auth"
+	"github.com/kubermatic/kubermatic/api/pkg/handler/v1/common"
 	"github.com/kubermatic/kubermatic/api/pkg/provider"
-	"github.com/kubermatic/kubermatic/api/pkg/version"
 )
-
-// UpdateManager specifies a set of methods to handle cluster versions & updates
-type UpdateManager interface {
-	GetVersion(string) (*version.MasterVersion, error)
-	GetMasterVersions() ([]*version.MasterVersion, error)
-	GetDefault() (*version.MasterVersion, error)
-	AutomaticUpdate(from string) (*version.MasterVersion, error)
-	GetPossibleUpdates(from string) ([]*version.MasterVersion, error)
-}
 
 // Routing represents an object which binds endpoints to http handlers.
 type Routing struct {
@@ -32,7 +23,7 @@ type Routing struct {
 	oidcAuthenticator     auth.OIDCAuthenticator
 	oidcIssuer            auth.OIDCIssuerVerifier
 	clusterProviders      map[string]provider.ClusterProvider
-	updateManager         UpdateManager
+	updateManager         common.UpdateManager
 	prometheusClient      prometheusapi.Client
 	projectMemberProvider provider.ProjectMemberProvider
 	userProjectMapper     provider.ProjectMemberMapper
@@ -48,7 +39,7 @@ func NewRouting(
 	projectProvider provider.ProjectProvider,
 	oidcAuthenticator auth.OIDCAuthenticator,
 	oidcIssuerVerifier auth.OIDCIssuerVerifier,
-	updateManager UpdateManager,
+	updateManager common.UpdateManager,
 	prometheusClient prometheusapi.Client,
 	projectMemberProvider provider.ProjectMemberProvider,
 	userProjectMapper provider.ProjectMemberMapper,
