+++ api/pkg/handler/test/helper.go
@@ -450,6 +450,34 @@ func GenTestMachine(name, rawProviderSpec string, labels map[string]string, owne
 	}
 }
 
+func GenTestMachineDeployment(name, rawProviderSpec string, selector map[string]string) *clusterv1alpha1.MachineDeployment {
+	var replicas int32 = 1
+	return &clusterv1alpha1.MachineDeployment{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: metav1.NamespaceSystem,
+		},
+		Spec: clusterv1alpha1.MachineDeploymentSpec{
+			Selector: metav1.LabelSelector{
+				MatchLabels: selector,
+			},
+			Replicas: &replicas,
+			Template: clusterv1alpha1.MachineTemplateSpec{
+				Spec: clusterv1alpha1.MachineSpec{
+					ProviderSpec: clusterv1alpha1.ProviderSpec{
+						Value: &runtime.RawExtension{
+							Raw: []byte(rawProviderSpec),
+						},
+					},
+					Versions: clusterv1alpha1.MachineVersionInfo{
+						Kubelet: "v9.9.9",
+					},
+				},
+			},
+		},
+	}
+}
+
 func CheckStatusCode(wantStatusCode int, recorder *httptest.ResponseRecorder, t *testing.T) {
 	t.Helper()
 	if recorder.Code != wantStatusCode {
